+++ packages/lila-core/src/app.js
@@ -3,7 +3,7 @@ import path from 'path';
 const { join } = path;
 
 export const cwd = process.cwd();
-export const tmpDir = join(cwd, '.lila');
+export const tmp = join(cwd, '.lila');
 
 export default {
   lila: undefined,