+++ packages/lila-webpack-config/src/build.js
@@ -14,10 +14,10 @@ export default (lila, webpack, { page, cmd, config }) => {
   const { getSettings } = lila;
   const [cwd, srcDir, buildDir, appDir, tmpDir] = getSettings([
     'cwd',
-    'srcDir',
-    'buildDir',
-    'appDir',
-    'tmpDir',
+    'src',
+    'build',
+    'app',
+    'tmp',
   ]);
   const realAppDir = join(cwd, appDir);
   const realSrcDir = join(realAppDir, srcDir);